Move player equips armor into a smart event
I'm going to hope @Morphan1 will handle all the TODO's I added
